requirements.txt
black==22.6.0
click==8.1.2
Flask==2.1.1
Frozen-Flask==0.18
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
Markdown==3.3.6
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
Pygments==2.7.1
PyYAML==5.1
tomli==2.0.1
Werkzeug==2.1.1
zipp==3.8.0
